+++ src/main/java/wtf/moneymod/client/impl/module/movement/Warp.java
@@ -12,7 +12,7 @@ public class Warp extends Module {
     //Pro module
 
     @Setting(id = "Time", clamp = @Clamp(min = 1, max = 16)) public int time = 10;
-    @Setting(id = "Tick", clamp = @Clamp(min = 1, max = 8)) public int tick = 4;
+    @Setting(id = "Tick", clamp = @Clamp(min = 1, max = 8)) public double tick = 4;
     @Setting(id = "Mode") public Mode mode = Mode.TIMER;
     @Setting(id = "Step") public boolean step = true;
     //pov when you not pasting ^
@@ -30,7 +30,7 @@ public class Warp extends Module {
         if (step) mc.player.stepHeight = 2;
         if (mode == Mode.TIMER) {
             delay++;
-            Main.TICK_TIMER = tick;
+            Main.TICK_TIMER = ( float ) tick;
             if (delay >= time) {
                 delay = 0;
                 setToggled(false);
